Added more functionality to the MindwaveMenu class
Added the class files for MindwaveMenu and ThinkGear.

Updated MindwaveMenu to grab data from the ThinkGear using the Happy
button, such as the waves and their values.

Added functionality to get new connection ID from the  Mindwave device
using the Thinkgear.class and the button in the MindwaveMenu.class
titled, Connect to device.

Setup two booleans to start and stop data recording.